COMP: remove warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=3745
git-svn-id: http://svn.slicer.org/Slicer4/trunk@3745 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Tractography/Display/vtkSlicerDiffusionTensorGlyphDisplayWidget.cxx
@@ -224,7 +224,7 @@ void vtkSlicerDiffusionTensorGlyphDisplayWidget::ProcessWidgetEvents ( vtkObject
         event == vtkKWScale::ScaleValueChangedEvent)
     {
     vtkErrorWithObjectMacro(this,"Process WIDGET... Events, number of tube sides set!  ;)");
-    displayNode->SetTubeGlyphNumberOfSides(this->TubeNumberOfSidesScale->GetWidget()->GetValue());
+    displayNode->SetTubeGlyphNumberOfSides((int) this->TubeNumberOfSidesScale->GetWidget()->GetValue());
     return;
     }
 }
@@ -374,7 +374,7 @@ void vtkSlicerDiffusionTensorGlyphDisplayWidget::UpdateMRML()
       displayNode->SetColorGlyphBy(this->GlyphColorMap[std::string(this->GlyphColorMenu->GetWidget()->GetValue())]);
       displayNode->SetGlyphEigenvector(this->GlyphEigenvectorMap[std::string(this->LineGlyphEigenvectorMenu->GetWidget()->GetValue())]);
       displayNode->SetGlyphScaleFactor(this->GlyphScale->GetWidget()->GetValue());
-      displayNode->SetTubeGlyphNumberOfSides(this->TubeNumberOfSidesScale->GetWidget()->GetValue());
+      displayNode->SetTubeGlyphNumberOfSides((int) this->TubeNumberOfSidesScale->GetWidget()->GetValue());
       }
     else 
       {
